build: streamline development and build pipelines
This commit does a massive rework of the entire project. It streamlines the processes to build, watch and develop the website.

This commit introduces the project as an NPM project, allowing us to use NPM packages and use its massive library of scripts.

We also use posthtml for our HTML development which allows us to minifiy theHTML and also use modules instead of duplicating code.

We introduce a live server to automatically watch and build the code as you go therefore having live changes