build: streamline development and build pipelines (#31)
* build: streamline development and build pipelines

This commit does a massive rework of the entire project. It streamlines the processes to build, watch and develop the website.

This commit introduces the project as an NPM project, allowing us to use NPM packages and use its massive library of scripts.

We also use posthtml for our HTML development which allows us to minifiy theHTML and also use modules instead of duplicating code.

We introduce a live server to automatically watch and build the code as you go therefore having live changes

* refactor: use modular code and cleanup

This commit deletes local versions of FontAwesome and Bootstrap since they make project size huge for no reason.

This commit also introduces `master.css` that imports `nav.css` for it to be on every page. And every other page imports it's own css file. `master.css` also imports bootstrap icons so we don't have to include it on each file.

This commit also uses modular HTML where we import `nav.html` on each file rather than duplicating code.

This commit also comments out some code that was causing an issue.